pyproject.toml
[tool.poetry]
name = "habbits-tracker"
version = "0.1.0"
description = ""
authors = ["denanmax <deshis93@gmail.com>"]
readme = "README.md"
packages = [{include = "habbits_tracker"}]

[tool.poetry.dependencies]
python = "^3.11"
django-restframework = "^0.0.1"
django = "^4.2.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
celery = "^5.3.4"
telebot = "^0.0.5"
redis = "^5.0.1"
pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
django-celery-beat = "^2.5.0"
telegram = "^0.0.1"
pytelegrambotapi = "^4.14.0"
coverage = "^7.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
